refactored process file to service
